Fix pdf copy for offline package to test for pdf first

+++ .teamcity/settings.kts
@@ -4347,7 +4347,7 @@ object GwBuildSteps {
                 #!/bin/bash
                 set -xe
                 
-                cp -avR "${onlineOutputPath}/pdf" "$offlineOutputPath"
+                [[ -e ${onlineOutputPath}/pdf/*.pdf ]] && cp -avR "${onlineOutputPath}/pdf" "$offlineOutputPath"
                 """.trimIndent()
         }
     }
